+++ GUI-dev/src/app/@theme/providers/backend-server/gateway.ts
@@ -1,7 +1,7 @@
 import {HttpClient, HttpParams} from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import {ServerConfigProvider} from './serverconfig';
-import {User, StoreProcess} from '../../../../models/models';
+import {User, StoreProcess, StoreProcessRating} from '../../../../models/models';
 
 
 @Injectable()
@@ -36,4 +36,15 @@ export class GatewayProvider {
       .toPromise()
   }
 
+  getStoreProcessRatings(processId: string): Promise<StoreProcessRating[]> {
+    const params = new HttpParams();
+    params.append('processId', processId)
+    return this.http.get<StoreProcessRating[]>(this.serverConfig.getStoreProcessRatings, { params: params })
+      .toPromise()
+  }
+
+  postStoreProcessRatings(processId: string, rating: StoreProcessRating): void {
+    const url = this.serverConfig.postStoreProcessRating + '/' + processId
+    this.http.post<StoreProcessRating>(url, rating).toPromise()
+  }
 }