+++ GUI-dev/src/app/@theme/providers/backend-server/serverconfig.ts
@@ -8,9 +8,12 @@ export class ServerConfigProvider {
   private _checkIfMailTaken = `${this._host}user/register/checkIfMailTaken/`;
   private _getUser = `${this._host}/api/me/`;
   private _getStoreProcesses = `${this._host}api/store/processes`;
+  private _getOrPostStoreProcessRatings = `${this._host}api/store/processRatings`;
 
   public get checkIfMailTaken(): string {return this._checkIfMailTaken};
   public get getUser(): string {return this._getUser};
   public get getStoreProcesses(): string {return this._getStoreProcesses};
+  public get getStoreProcessRatings(): string {return this._getOrPostStoreProcessRatings};
+  public get postStoreProcessRating(): string {return this._getOrPostStoreProcessRatings};
 
 }