+++ GUI-dev/src/app/pages/pages.module.ts
@@ -9,6 +9,8 @@ import {ApprovalModule} from './approval/approval.module';
 import {ApprovalDetailsModule} from './approval-details/approval-details.module';
 import AdminModule from './admin/admin.module';
 import {UserDetailsModule} from './user-details/user-details.module';
+import { ProcessRatingModule } from './includes/process-ratings/process-ratings.module';
+import { ProcessstoreDetailsModule } from './processstore-details/processstore-details.module';
 
 const PAGES_COMPONENTS = [
   PagesComponent,
@@ -25,6 +27,8 @@ const PAGES_COMPONENTS = [
     AdminModule,
     UserDetailsModule,
     ProcessStoreSearchModule,
+    ProcessRatingModule,
+    ProcessstoreDetailsModule,
   ],
   declarations: [
     ...PAGES_COMPONENTS,